Optionally import torch and enable Proxy type in Numbast
